+++ app/web/features/analytics/eventCollector.ts
