+++ app/web/features/analytics/index.ts
@@ -1,9 +1,9 @@
-export { logEvent } from "./event-collector";
+export { logEvent } from "./eventCollector";
 export { useDurationEvent, useImpressionRef, useLogEvent } from "./hooks";
 export {
   useFormInteraction,
   useFunnelStep,
   useHoverDwell,
   useScrollDepth,
-} from "./journey-hooks";
+} from "./journeyHooks";
 export { default as AnalyticsProvider } from "./provider";
